muscle/pyproject.toml
[project]
name = "muscle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "gspread>=6.2.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "plotly>=6.6.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
